Modernise
Generalised script for numerous cores.

Only concern is that each swap 'file' is allocated the wrong maximum amount of memory; couldn't find best-practice.
As the compression is more than one-to-one, the compressed entity will be smaller than the original, so we'll always have more space than before the swap existed.

Not useful for a Pi zero as is. Strip out the guts to one line.